IIT REVIN/requirements.txt
pymupdf==1.24.0
langchain==0.2.0
langchain-google-genai==1.0.0
langchain-community==0.2.0
langchain-chroma==0.1.0
chromadb==0.5.0
streamlit==1.35.0
python-dotenv==1.0.0
google-generativeai==0.7.0
Pillow==10.3.0
groq